# marketing_op/api/design/openapi.yaml
openapi: 3.1.0
info:
  title: Marketing Data (Objective Platform)
  version: 1.0.0
  description: "API for fetching marketing data measurements and insights."
paths:
  /api/auth/token/:
    get:
      operationId: api_marketing_op_api_get_token
      summary: Get Token
      description: Retrieves the token necessary to authenticate the rest of the endpoints of the API.
      security:
        - BasicAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    properties:
                      token:
                        type: string
                        example: "eyJhbGciOi"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid credentials"
  /api/marketing/measurements/:
    get:
      operationId: marketing-measurements
      summary: Get Marketing Measurements
      security:
        - AuthBearer: []
      parameters:
      - in: query
        name: products
        schema:
          type: array
          items:
            type: string
          title: Products
        required: false
      - in: query
        name: campaigns
        schema:
          type: array
          items:
              type: string
        required: false
      - in: query
        name: campaign_types
        schema:
          type: array
          items:
            type: string
            enum:
              - always on
              - branding
              - performance
        required: false
      - in: query
        name: channels
        schema:
         type: array
         items:
          type: string
          enum:
            - tv
            - radio
            - ooh
            - facebook
            - print
            - tiktok
            - instagram
            - db360 - prospecting
            - db360 - retargeting
            - google - branded search
            - google - non-branded search
            - dv360 - prospecting
            - dv360 - retargeting
            - folder
        required: false
      - in: query
        name: start_date
        schema:
          format: date
          type: string
        required: false
      - in: query
        name: end_date
        schema:
          type: string
          format: date
        required: false
      - in: query
        name: offset
        schema:
          default: 0
          minimum: 0
          type: integer
        required: false
      - in: query
        name: page_size
        schema:
          default: 10
          exclusiveMinimum: 0
          maximum: 100
          type: integer
        required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      $ref: "#/components/schemas/MarketingDataSchema"
                    title: Data
                    type: array
                    total: 1
                required:
                  - data
              examples:
                MarketingData:
                  summary: Marketing Data
                  value:
                    data:
                      - product: "Product A"
                        campaign_type: "always on"
                        campaign: "launch campaign new apple brand"
                        channel: "tv"
                        date: "2021-01-01"
                        conversions: 1225.0170852049143
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "End date should be greater than start date"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Invalid token"
  /api/marketing/stats/channel-sales-percentages/:
    get:
      operationId: api_marketing_op_api_channel_sales_percentages
      summary: Channel Sales Percentages
      security:
        - AuthBearer: []
      parameters:
        - in: query
          name: products
          schema:
            type: array
            items:
              type: string
          required: false
        - in: query
          name: campaigns
          schema:
            type: array
            items:
              type: string
          required: false
        - in: query
          name: campaign_types
          schema:
            type: array
            items:
              type: string
              enum:
                - always on
                - branding
                - performance
          required: false
        - in: query
          name: channels
          schema:
            type: array
            items:
              type: string
              enum:
                - tv
                - radio
                - ooh
                - facebook
                - print
                - tiktok
                - instagram
                - db360 - prospecting
                - db360 - retargeting
                - google - branded search
                - google - non-branded search
                - dv360 - prospecting
                - dv360 - retargeting
                - folder
          required: false
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          example: "2021-01-01"
          required: false
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          example: "2021-05-01"
          required: false
        - in: query
          name: offset
          schema:
            default: 0
            minimum: 0
            type: integer
          required: false
        - in: query
          name: page_size
          schema:
            default: 10
            exclusiveMinimum: 0
            maximum: 100
            type: integer
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      $ref: "#/components/schemas/ChannelSalesPercentageSchema"
                    type: array
                required:
                  - data
              examples:
                ChannelSalesPercentage:
                  summary: Channel Sales Percentage
                  value:
                    data:
                      - channel: "tv"
                        percentage: "10.00"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
  /api/marketing/stats/channel-weekly-sales/:
    get:
      operationId: api_marketing_op_api_channel_weekly_sales
      summary: Channel Weekly Sales
      security:
        - AuthBearer: []
      parameters:
        - in: query
          name: products
          schema:
            type: array
            items:
              type: string
          required: false
        - in: query
          name: campaigns
          schema:
            type: array
            items:
             type: string
          required: false
        - in: query
          name: campaign_types
          schema:
            type: array
            items:
             type: string
             enum:
               - always on
               - branding
               - performance
          required: false
        - in: query
          name: channels
          schema:
            type: array
            items:
             type: string
             enum:
               - tv
               - radio
               - ooh
               - facebook
               - print
               - tiktok
               - instagram
               - db360 - prospecting
               - db360 - retargeting
               - google - branded search
               - google - non-branded search
               - dv360 - prospecting
               - dv360 - retargeting
               - folder
          required: false
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: offset
          schema:
            default: 0
            minimum: 0
            type: integer
          required: false
        - in: query
          name: page_size
          schema:
            default: 10
            exclusiveMinimum: 0
            maximum: 100
            type: integer
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data:
                    items:
                      $ref: "#/components/schemas/ChannelWeeklySalesSchema"
                    title: Data
                    type: array
                    total: 1
                required:
                  - data
              examples:
                ChannelWeeklySales:
                  summary: Channel Weekly Sales
                  value:
                    data:
                      - channel: "tv"
                        year: 2024
                        week: 3
                        sales: 1225.0170852049143
components:
  schemas:
    MarketingDataSchema:
      properties:
        product:
         type: string
         example: "Product A"
        campaign_type:
          type: string
          example: "always on"
          enum:
            - always on
            - branding
            - performance
        campaign:
          type: string
          example: "launch campaign new apple brand"
        channel:
          type: string
          enum:
            - tv
            - radio
            - ooh
            - facebook
            - print
            - tiktok
            - instagram
            - db360 - prospecting
            - db360 - retargeting
            - google - branded search
            - google - non-branded search
            - dv360 - prospecting
            - dv360 - retargeting
            - folder
          example: "tv"
        date:
          type: string
          format: date
          example: "2021-01-01"
        conversions:
          type: number
          example: 1225.0170852049143
      title: MarketingDataSchema
      type: object
    ChannelSalesPercentageSchema:
      properties:
        channel:
          type: string
          enum:
            - tv
            - radio
            - ooh
            - facebook
            - print
            - tiktok
            - instagram
            - db360 - prospecting
            - db360 - retargeting
            - google - branded search
            - google - non-branded search
            - dv360 - prospecting
            - dv360 - retargeting
            - folder
        percentage:
          type: string
          example: "10.00"
      required:
        - channel
        - percentage
      type: object
    ChannelWeeklySalesSchema:
      properties:
        channel:
          type: string
          enum:
            - tv
            - radio
            - ooh
            - facebook
            - print
            - tiktok
            - instagram
            - db360 - prospecting
            - db360 - retargeting
            - google - branded search
            - google - non-branded search
            - dv360 - prospecting
            - dv360 - retargeting
            - folder
        year:
          type: integer
          example: 2024
        week:
          type: integer
          example: 3
        sales:
          type: number
          example: 1225.0170852049143
      required:
        - channel
        - year
        - week
        - sales
      type: object
  securitySchemes:
    AuthBearer:
      type: http
      scheme: bearer
    BasicAuth:
      type: http
      scheme: basic
servers: []